backend/requirements.txt
flask
flask-cors
graphene
flask-jwt-extended
Authlib
Flask-JWT-Extended
graphql-server
mongoengine
pymongo
python-dotenv
mongomock
flask_jwt_extended
mailgun2
validate_email_address
pika
flask-socketio
redis
gevent
gevent-websocket
gunicorn
dill
cryptography
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-instrumentation-pymongo
inflection
